Refactored time module into separate files for better readability.
