+++ src/utils/time/index.ts
@@ -1,1 +1,3 @@
+export * from './getTimeUntil';
 export * from './time';
+export * from './timeTypes';
